Use portal for Modal
Currently Modals are susceptible to being mis-positioned when they are instantiated from within a DOM element that is positioned absolute or fixed.

The ModalContainer now uses a React.Portal to inject itself outside of the existing DOM structure to ensure that it is not effected by its mount location.

Portals were literally built for this use case.

In the interest of keeping the implementation easy to use the Modal component will automatically create a DOM element at the end of the window.document to hook to. It will automatically re-use that that some element if it’s mounted again in the future.