Menu component refinements
MenuItem should not be cast to an `a` element (`is=a`) because it makes composition with Lens’ `<Link>` as well as ReactRouter’s `<Link>` component unnecessarily complex. It also is technically an inaccurate semantic definition in many cases.

 Changing MenuItem m* & p* to smooth out scrolling behaviors when inside a MenuGroup with a label assigned.

Adjusted styling of MenuGroup label to look better when scrolling. Still needs further refinement in cases where menu doesn’t require scrolling (boxShadow feels superflous in those cases)

Updated