Icon component support for innerRef property
The updated implementation of Overlay expects to be able to directly assign the `innerRef` property to the trigger (children). Icon's current implemntation (Box wrapped in styled() call) prevents this direct association.

There are two places in Lens where Icon is "extended" to apply styles to it within the styles of another component so for those this commit extends Button in a context specific to that component to allow its reference within the component specification.