Don't subclass ModalHoverManager
Subclassing is an antipattern in React, and typescript is particularly ill-equipped to handle the prop types of subclassed React components. This is the minimal commit to remove subclassing from Tooltip and RichTooltip. We change renderModal from a virtual method with overrides to a lambda prop, and pass in state and instance variables as arguments.