Remove server package
With the upcoming move to consolidate monorepos the server package will
likely be replaced with Looker Extensions for integration testing so it
no longer makes sense to maintain this package.

If, at a future point it makes sense to restore this package – it can be
pulled back from git history.