fix: Correct web site publishing to work with lerna independent versioning (#2514)
This PR corrects a documentation web site deploy bug where `lerna.json` `version` key is still leveraged (it's now "independent"). Current each version of the site overwrites into a `vindependent` folder instead of `v1.1.1` 

`latest` which is the publicly visible version still works fine but it'd be ideal to straighten this out so we can keep collecting "historic" versions of the documentation site for easy reference.