chore: allow gatsby mdx files to be used directly in developer portal (#2820)
The Looker developer portal uses the gatsby mdx files for documenting components. This change allows the mdx files to be used without modification in both the Looker Components gatsby build and the Looker developer portal gatsby build.

The fix is to add shortcodes to the MDXProvider. MDX files no longer import local classes directly and instead get them from the MDXProvider. It is okay to import components from npm packages but not if they are deep imports. Note that build issues occur if functions are exported as shortcodes. As such the use colorBreakdown has been moved from color.mdx into the ColorsList.tsx and ColorBlendsGrid.tsx.

The colorBreakdown function and associated types are now exported from the design components package.